dom/src/threads/test/test_json.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker JSON messages
-->
<head>
  <title>Test for DOM Worker Navigator</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script src="json_worker.js" language="javascript"></script>
<script class="testbody" language="javascript">

  ok(messages.length, "No messages to test!");

  var worker = new Worker("json_worker.js");

  var index = 0;
  worker.onmessage = function(event) {
    var key = messages[index++];

    // Loop for the ones we shouldn't receive.
    while (key.exception) {
      key = messages[index++];
    }

    is(typeof event.data, key.type,
       "Bad type! " + messages.indexOf(key));
    is(event.data instanceof Array, key.array,
       "Array mismatch! " + messages.indexOf(key));

    if (key.shouldCompare) {
      ok(event.data == key.compareValue,
         "Values don't compare! "  + messages.indexOf(key));
    }

    if (key.shouldEqual) {
      ok(event.data === key.compareValue,
         "Values don't equal!" + messages.indexOf(key));
    }

    if (event.data == "testFinished") {
      is(index, messages.length, "Didn't see the right number of messages!");
      SimpleTest.finish();
    }
  };

  worker.onerror = function(event) {
    ok(false, "Worker had an error: " + event.data);
    SimpleTest.finish();
  }

  worker.postMessage("start");

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
